requirements.txt
requests>=2.0,<3.0
python-dotenv>=1.0,<2.0
